fix populatesensordata

+++ models/sensordata.go
@@ -8,7 +8,7 @@ type SensorData struct {
 	TDS                 float32   `json:"tds"`
 	PH                  float32   `json:"ph"`
 	SolutionTemperature float32   `json:"solution_temp"`
-	GrowUnitLux         []float32 `json:"grow_unit_lux"`
-	GrowUnitHumidity    []float32 `json:"grow_unit_humidity"`
-	GrowUnitTemperature []float32 `json:"grow_unit_temp"`
+	GrowUnitLux         []float64 `json:"grow_unit_lux"`
+	GrowUnitHumidity    []float64 `json:"grow_unit_humidity"`
+	GrowUnitTemperature []float64 `json:"grow_unit_temp"`
 }
